crawltikitrends/requirements.txt
attrs==25.3.0
Automat==25.4.16
awscli==1.40.35
botocore==1.38.36
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.4
cssselect==1.3.0
defusedxml==0.7.1
docutils==0.19
filelock==3.18.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.4.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
parsel==1.10.0
Protego==0.4.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2025.2
PyYAML==6.0.2
queuelib==1.8.0
requests==2.32.4
requests-file==2.1.0
rsa==4.7.2
s3transfer==0.13.0
Scrapy==2.13.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
tldextract==5.3.0
Twisted==25.5.0
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
w3lib==2.3.1
zope.interface==7.2
